pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdrive-cleaner"
version = "0.1.0"
description = "Google Drive maintenance CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-api-python-client",
  "google-auth",
  "pandas>=2.2",
  "openpyxl",
  "python-calamine>=0.2.0",
  "rich",
  "tenacity",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "python-dotenv>=1.0",
  "pytest-mock>=3.0",
]
notebook = [
  "ipykernel>=6.29",
  "jupyter>=1.0",
  "notebook>=7.0",
]

[project.scripts]
gdrive-cleaner = "gdrive_cleaner.cli:main"
gdc = "gdrive_cleaner.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
